Added functionality for list and layout for update a 3dprint

+++ Code/3DMANAGER-APP/3DMANAGER-APP.BLL/Models/Filament/FilamentRequest.cs
@@ -8,7 +8,7 @@
         public decimal FilamentWeight { get; set; }
         public string FilamentColor { get; set; }
         public int FilamentTemperature { get; set; }
-        public int FilamentLenght { get; set; }
+        public decimal FilamentLenght { get; set; }
         public decimal FilamentThickness { get; set; }
         public decimal FilamentCost { get; set; }
         public string FilamentDescription { get; set; }
